Fix Physical Tile Counter
Currently, the total number of physical tiles read does not match the sum of
the fixed-size and var-sized tiles read if more than one var-sized tile is read.
For example:
```
- Number of physical tiles read: 833
  * Number of physical fixed-sized tiles read: 441
  * Number of physical var-sized tiles read: 784
```

This patch fixes the issue:
```
- Number of physical tiles read: 1225
  * Number of physical fixed-sized tiles read: 441
  * Number of physical var-sized tiles read: 784
```

This also removes the READ_ATTR_NUM and READ_DIM_NUM stats because they are
redundant.