Lazily Init AWS ClientConfiguration
AWS changes the ClientConfiguration in the 1.8 SDK to do the checking for env
variables and ec2 metadata in [1]. This can cause TileDB to behavior slow if
S3 support is built but the environment is not configured. The AWS SDK
check for the ec2 metadata and has to wait for a timeout. We need to
lazily init the ClientConfiguration.

[1] https://github.com/aws/aws-sdk-cpp/commit/147469373c9fec1037bd2d75d7cd949250c6f7c5