Writer should serialize Subarray object
This fixes a problem with serialized queries for doing dense writes with
ranges. Previously dense writes only worked with subarrays but not
ranges. The user would receive an error about not enough cells being
written as without any ranges the dense write would assume the user was
writting the entire domain of the array.