Adjust fill in value support for issue with REST
The fill in value is only serialized to disk if the array schema version
is`>= 6`. Previously if the schema was created from de-serializing from
Cap'n Proto and it was a version 5 or older then the fill in values were
incorrectly serialized on disk causing a corrupt array schema that could
not be read.