LRUCache::insert() bugs
This fixes two long-standing bugs our LRU implementation, which is shared between
the tile cache and the new VFS read-ahead cache.

1. When inserting an element, we evict at least one object in the cache to
make room for the new object. If we evict an object with the same key as the
object being inserted, the `item_it` iterator is invalidated and segfaults
when used.

2. The `LRUCache::size_` corrupts every time that an object is inserted that
shares a key with an existing object.