Disable tiledb::sm::Domain copy-constructor (#1885)
Currently, the `tiledb::sm::Domain` class has a default copy-constructor. The
default copy-constructor performs a copy of unmanaged `Dimension` pointers. After
the copy, it ambigious whether the original or the copy have ownership of the
`Dimension` pointers. The `tiledb::dm::Domain` destructor frees the the unmanaged
pointers, causing either a double-free or use-after-free scenario.

This issue currently manifests within the `unit-rtree.cc` where a `Domain` instance
is returned from the `Domain create_domain()` routine. If the compiler does not
perform return value optimization (RVO) to elide the copy, it returns a shallow
copy.

This patch disables the copy-constructor and copy-assignment operators for the
`tiledb::sm::Domain`. It introduces move-constructor and move-assignment operators
for the compiler to use when it does not perform RVO.

Additionally, this includes a few other related/misc changes:
- The unmanaged `Dimension` pointers are now `unique_ptr<Dimension>` objects.
- The copy-constructor and copy-assignment operators are disabled for the
  `tiledb::sm::Attribute` and `tiledb::sm::Dimension` classes.

Co-authored-by: Joe Maley <joe@tiledb.com>