Always serialize ArraySchema with the current format version
This fixes a bug in the following scenario:
1. Read an array with an older format version
2. Write an a new array, using the array schema from the previous step.
3. Read the new array.

Note that our serialization path does not check the version. We do not support
writing array schemas as if they were being written from an earlier version
of the core.

In step #2, we serialize the array schema as if it was the latest version but
set the "version" attribute to an older version.
In step #3, we read the "version" attribute and deserialize that format. The
format is a newer format, so the deserialization fails.

---

There are a few options to fix the bug:
1. Always write the "version" as the latest version. This ensures future reads
can correctly deserialize the on-disk format.

2. Treat all deserialized objects (e.g. in-memory ArraySchema objects) as if
they were the latest version.

3. Modify the serialization path for backwards compatibility.

Solution 3 is a lot of work, and I'm not sure if we even need/want to support
that use-case. Solution 2 is my preferred fix, but there may be a future use-case
where we want to know that version of an ArraySchema after it has been deserialized.
This patch implements solution 1.