MemFS node-level locks
This removes the instance-wide mutex in the MemFilesystem class for node-level
mutexes.

The primary motiviation is to allow concurrent IO on different node instances.
I've also included some misc style changes and clean-up that I missed in the
initial review.

The locking rule for the directory tree is that we always lock top-down. This
ensures that we will not deadlock.