Handle `must_split_slab` in SubarrayPartitioner::split_current()
After calibrating `current_` in `SubarrayPartitioner::split_current()`, the
`must_split_slab` may be true. In this scenario, the reader will get stuck
in an infinite loop because it can not make progress on the current partition.

This change handles `must_split_slab == true` to split the current partition.
I verified the fix in a repro scenario that was previously hanging. It now
completes with the expected number of records.