Using our OpenSSL from superbuild requires pkg-config (#1927)
When we build OpenSSL from source, we tell the Curl build to use OpenSSL's
pkg-config files to find OpenSSL. If pkg-config is not available, the
curl configure fails beacuse the test programs won't build. This is
buried inside of Curl's config.log and the problem is not directly
obvious. The Curl configure fails with a very unhelpful error, and
the underlying reason is the test programs try to link default
lib/include paths rather than our 'externals/install/lib'.
At best the top level build will fail with a confusing error,
at worst we might link a system SSL we don't want in superbuild
mode.

Co-authored-by: joe maley <joe@tiledb.io>