Fix consolidation using incorrect fragment size
We now store the fragment footer size on loading of the fragment so that
is can be referenced later without recomputing or re-reading of the
footer. This allows supporting string dimensions, as the footer size is
variable length and stored at the end of the footer. Previously the call
to FragmentMetadata::get_footer_size only accounted for
fixed size fragment footers, and doesn't attempt to read the size at the
end of the footer as is required for string dimensions.