Parallelize Writer::write_tiles
Within `Writer::write_all_tiles` we parallelize among attributes, but we do
not parallelize among the data tiles. This patch parallelizes writes among
data tiles, as well as the among fixed/var/null files.

In my testing on a 4D array with a single fixed attribute, the write tiles
execution time improved by 40% on an S3 backend. However, this patch is
applicable to any backend.