Modify C++ API Query::est_result_size_var (#1946)
This changes the return type and return values for `Query::est_result_size_var`.

Currently, the size estimate for the values are returned in bytes while the size
estimate for the offsets are returned in elements. For consistency, this API
now returns both the size estimates as bytes. This is breaking behavior.

To make the breaking behavior more obvious, we have modified the return type
from an `std::pair` to an `std::array`.

We chose to make both of these size estimates return as bytes rather than
elements to stay consistent with the existing `Query::est_result_size` for
fixed-size attributes.

Co-authored-by: Joe Maley <joe@tiledb.com>