Fix GCS read-ahead
This fixes the following issue reported in TileDB-Py:
https://github.com/TileDB-Inc/TileDB-Py/issues/443
"TileDBError: [TileDB::?] Error:: Read object failed on: gcs://tiledb/test/__array_schema.tdb"

The issue is that if the `stream` in GCS::read is read beyond the length of
the stream, the `operator!` will return true. I'm not sure whether this is a
bug GCS client library or not. I would only expect the stream to reach the EOF
rather than a failed state.

Either way, we do our own error-checking later in this path by comparing
`if (*length_returned < length)`. This patch removes the check on `!stream`.

---

With our recent VFS refactor in the test path, I was easily able to connect test
GCS before-and-after this patch. Without this patch, many unit tests fail.
With the patch, all of the unit tests pass. This is important to note because
GCS is not under test in the CI -- if it was, we would have caught this regression
much earlier.

Since I've already written the SupportedFsGCS class for testing purposes, I've
included in this patch as well. Note that I've configured the GCS project id
to "TODO", which will be changed when we include the GCS emulator in CI. For
my testing purposes, I had replaced "TODO" with the project-id on the real GCS.