Fix GCS read-ahead (#1972)
This fixes the following issue reported in TileDB-Py:
TileDB-Inc/TileDB-Py#443
"TileDBError: [TileDB::?] Error:: Read object failed on: gcs://tiledb/test/__array_schema.tdb"

The issue is that if the stream in GCS::read is read beyond the length of
the stream, the operator! will return true. I'm not sure whether this is a
bug GCS client library or not. I would only expect the stream to reach the EOF
rather than a failed state.

Either way, we do our own error-checking later in this path by comparing
if (*length_returned < length). This patch removes the check on !stream.

Co-authored-by: Joe Maley <joe@tiledb.com>