Optimize consolidated fragment metadata loading
This is part one of two changes for optimizing consolidated fragment
metadata loading. This changes removes using VFS to always fetch the
metadata file size. Instead we select the file size only if not reading
from a consolidated file buffer. We also adjust the `fragment_size`
function, used by consolidation, to fetch the fragment metadata file
size on request. The fetching of the size on-demand for consolidation
will yield the same performance degredation we see in open array,
however this is acceptable for patch one and will be addressed in the
next series with a format change.