Optimize load_tile_offsets for only relevant frags
This optimization is to adjust the `Reader::load_tile_offsets` class to
only loop over the relevant fragments as computed by the subarray class
based on intersection. This is a performance optimization for arrays
which have a large number of fragments and which we are incorrectly
fetching a large amount of unneeded data.