Add support for s3 requester pays
If an AWS bucket has the requester pays option enabled a separate
parameter is needed by the client to signal to AWS that the user accepts
responsiblity for the s3 access charges. This change adds a new boolean config
option `aws.s3.requester_pays` to allow the user to set this behavior.