Zero-copy buffers in ArrowAdapter
- Added a bootstrap flag to enable the arrow unit tests. Note that I changed
the name of the env variable from `TILEDB_TESTS_ENABLE_ARROW` to
`TILEDB_ARROW_TESTS` so that the env var name matches the name of the bootstrap
flag.

- The `ArrowAdapter` now assumes that the core is configured as follows, allowing
zero-copy in both the import (write-to-tiledb) and export (read-from-tiledb)
path:
```
config["sm.var_offsets.bitsize"] = 32;
config["sm.var_offsets.mode"] = "elements";
config["sm.var_offsets.extra_element"] = "true";
```

- As outlined in-code, there is a bug in the Query::get_buffer() API when the
config "sm.var_offsets.bitsize" is 32. As a temporary work-around, we can
invoke the `tiledb_query_get_buffer_var` C-API.

- The work-around requires passing the C++ API `Context` object into the
`ArrowAdapter`. For consistency, I was going to pass this as a `shared_ptr`
like the existing `Query` argument. It does not look like it is necessary
for the `Query` to be a `shared_ptr` so I made both the `ctx` and `query` objects
raw pointers.