Percent-encode reserved characters in file names
This bumps the format version from 7 to 8. For version 8, dimension and
attribute names are percent-encoded to RFC 3986 + a few extra to handle
reserved characters on Windows.

This is intended to be a temporary measure. In the future, we will include
a mapping between names and their ordered index in the array schema.