Release s3 multipart lock early in flush_object
For S3 with mutipart uploads, the last step is to flush the object. We
grab an exclusive lock to get the shared state, but don't unlock it
until after we complete the upload. There is no need to hold the lock
after we get the shared state. Unlocking early will yield better
parallelism and performance.