Automate HISTORY.md
This adds a new github action that will scan the PR description for lines that
start with the following keywords: "TYPE:" and "DESC:". The action requires
at least one pair of these keywords to be present. There may be more than
one pair.

For each pair, a line will be added to the HISTORY.md under the header that
corresponds to the type. The line will be formatted with the associated
description. There is also a reserved keyword that can be used to bypass the
bot the in event that a PR wants to opt out of updating the history. This
keyword can appear anywhere in the PR description.

When the action updates the HISTORY.md, it will add a new commit to the
associated branch. If the action is triggered but the annotations in the PR
are unchanged, the action detects this and omits a new commit.

---

TYPE: IMPROVEMENT
DESC: Adds a Github Action to automate the HISTORY.md