CI: Heap memory API violations (#2070)
This introduces a new script, scripts/find_heap_api_violations.py, which is
a best-effort to detect usage of standard C/C++ dynamic memory APIs. The
intent is to catch scenarios where a new PR does not use the new APIs defined
in tiledb/common/heap_memory.h. We want all dynamic memory to happen through
these APIs so that we can record them in the heap profiler.

Co-authored-by: Joe Maley <joe@tiledb.com>