Limit range splits on string dimensions
This patch introduces a "depth" attribute on a `Range` type. Consider the
scenario where the partitioner splits an int range {1, 4} as far as it can:
```
depth 0: {1, 4}
depth 1: {1, 2}, {3, 4}
depth 2: {1, 1}, {2, 2}, {3, 3}, {4, 4}
```

The partitioner can not split the ranges at depth 2 any further. Unlike integer
ranges, string ranges can be split indefinitely. Consider the scenario where
the partitioner splits a string range { { 97 }, { 98 } } (ascii range {"a",
"b"}) as far as it can:
```
depth 0: {{97}, {99}}
depth 1: {{97}, {98 127}}, {{99}, {99}}
depth 2: {{97}, {97 127}}, {{98}, {98 127}}, ...
depth 3: {{97}, {97 63 127}}, {{97 64}, {97 127}}, ...
...
depth 21: {{97 1 1}, {97 1 1 3 127}}, {{97 1 1 4}, {97 1 1 7 127}}, ...
```

In the partitioner, it may split a string dimension to find estimated results
that fit the budget. There is a scenario where estimated results will never
fit the budget, regardless of many times the string dimension split. This causes
an infinite loop. To fix this scenario, this patch introduces a constant to
define the maximum split depth of a string dimension.

A depth of D will limit the total number of splits to `2^D - 1`. It is set to `7`
to split a string dimension no more than `127` times.

---

TYPE: BUG
DESC: Fix an edge-case where a read query may hang on array with string dimensions