sm::array_open_for_reads: fetch fragments async instead of schema
Invert the tasking order for async loading established in
  https://github.com/TileDB-Inc/TileDB/pull/2061

Now load the fragments in a task and array schema on main.
Fixes several mutex ownership issues which surfaced on Windows.

(1) We need to wait on the async task in all return clauses to avoid
    the following situation which causes an exit on Windows:
    if the context owning the storage manager is destroyed
    without joining the async thread, then we get an error
    on Windows due to attempting to destruct a locked mutex.

(2) We cannot call open_array->mtx_unlock() on the main thread after
    running the array_open call async on the thread pool (which locks
    the open_array.mtx_), because lock/unlock on different threads
    is undefined behavior and a runtime error on windows.