Multiple attributes in the smoke test
This refactors the smoke test to test attributes ["a"] and ["a", "b"] rather
than each attribute individually. The motiviation is that this is a more
complex path and may uncover issues that testing single attributes may not. I
specifically want this because I intend to extend the smoke test for the
next feature that I'm working on (attribute conditionals).

---

TYPE: NO_HISTORY