Adjust serialized query buffer sizes (#2115)
Adjust serialized query buffer sizes

This change the client/server flow to always send the server the
original user requested buffer sizes. This solves a bug in which with
serialized queries incompletes would cause the "server" to use smaller
buffers for each iteration of the incomplete query. This yield
decreasing performance as the buffers approached zero. The fix here lets
the server always get the original user's buffer size.