Share `tile_overlap_` with partitioned `Subarray` instances.
Currently, a partitioned Subarray makes a copy of all relevant `tile_overlap_`
elements in the parent. This data structure may be large. This patch shares
the data structure to deduplicate memory and save on copy time.

This changes the contract to `Subarray::tile_overlap`. The caller is now
responsible for indexing the ranges in the return value as if they were the
range indexes in the parent.

Additionally, I've removed `Subarray:tile_overlap_computed_` because we can
derive the same information by checking if `Subarray:tile_overlap_` is a nullptr.

---

TYPE: IMPROVEMENT
DESC: Reduced memory consumption in the read path for multi-range reads.