Unsplittable serialized queries need to recompute tile_overlap (#2120)
* Unsplittable serialized queries need to recompute tile_overlap

If a query is unsplittable then when it is resubmitted the query does not
move to the next partition and instead attempts to run the current one
again. Serialization of queries does not ship the tile overlap since it
can be large. As such when deserializing if the read_state_ is
unsplittable we need to make sure we recompute the tile_overlap for the
current subarray so it is available for the read.

* Update HISTORY.md for #2120

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>