Fix bug in `ResultTile::compute_results_sparse<char>` (#2125)
* Fix bug in `ResultTile::compute_results_sparse<char>`

The compute_results_sparse partitions in the input coordinates into ranges.
The size of the last coordinate is incorrectly calculated, which may corrupt
the corresponding entry in `result_bitmap` for this coordinate.

To calculate the size of a coordinate, we take the difference between two
contiguous offsets. We do not have a trailing offset for the last coordinate,
so we must leverage the total buffer size. Currently, we incorrectly do this
for the last coordinate in _every_ partition. We need to do this for the last
coordinate in the _last_ partition.

---

TYPE: BUG
DESC: Corrected a bug where sparse cells may be incorrectly returned using string dimensions.

* Update HISTORY.md for #2125

Co-authored-by: Joe Maley <joe@tiledb.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>