S3 Server-Side Encryption (#2130)
* S3 Server-Side Encryption

Introduces config option "vfs.s3.sse". Valid non-empty values are "aes256" and
"kms". These are the two supported SSE types in the AWS SDK.

If non-empty, the S3 class will use SSE on all of the requests we use that
support SSE.

This was manually tested by creating an array with "aes256" and using the
S3 web portal to verify SSE is enabled on the objects.

---

TYPE: FEATURE
DESC: Config option "vfs.s3.sse" for S3 server-side encryption support

* Update HISTORY.md for #2130

Co-authored-by: Joe Maley <joe@tiledb.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>