Fix consolidated fragment metadata parsing (#2135)
* Fix consolidated fragment metadata parsing

Currently, all footers in consolidated metadata files are parsed with the format
version of the consolidated metadata file. This patch ensures that footers are
parsed with the version stored in the footer.

In practice, this can fix one potential error when opening an array:
TileDBError: [TileDB::ConstBuffer] Error: Read buffer overflow

---
TYPE: BUG
DESC: Fixes a potential crash when opening an array with consolidated fragment metadata

* Update HISTORY.md for #2135

Co-authored-by: Joe Maley <joe@tiledb.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>