Parallelize file_size lookup in get_fragment_info (#2143)
Parallelize file_size lookup in get_fragment_info

The call to `fragment_size` is expensive, it is also safe to do in
parallel. This shifts to calling fragment_size inside a parallel_for to
speed up this operation. In testing on an array with 2600 fragments, this
drops the loading of fragment metadata from 96s to 11s.