Limit `TileOverlap` instances while partitioning
The memory consumption from the tile overlap state scales by the number of
fragments and number of ranges. This patch allows the subarray partitioner to
split subarrays that it estimates would consume too much memory. The motivation
is that the split subarrays will have a smaller number of ranges and therefore
have a smaller memory consumption from the tile overlap state.

The introduces a config option, `sm.max_tile_overlap_size` that the partitioner
uses to limit the range of the subarray it is operating on. This config option
is a target and not a hard limit.

The C APIs that estimate result sizes path through the
`Subarray::compute_est_result_size` routine. This routine is not yet aware
that the `tile_overlap_` may not be truncated due to the memory constraint. This
path forces the computation to override the memory constraint for now.

---

TYPE: IMPROVEMENT
DESC: Reduces memory usage in multi-range range reads