Add c/c++ api for getting config from query
This introduces a new c api `tiledb_query_get_config` and a new C++ api
`Query.config()` which fetches the config that is set on the query. This
can either be a config from `set_config` or the ctx's config if the user
never set one on the query explicitly.