Add config option to bound read ranges to domain
This adds a new option `sm.read_range_oob` which defaults to "error".
When set to true any range that is out of bounds for the domain of an
array will cause an error to be returned and the range to not be set.
This is the existing TileDB behavior.

If the option is instead set to "warn", the behavior changes and we print
a warning about the out of bounds but instead mutate the range so the
lower/upper bound is equal to the domain. This improves the user
experience greatly by allowing a user who does not know the domain to
issue a broad range and get results returned.