Fix memory leak in filter pipeline
In some scenarios, the "filter_chunks_reverse" is reached with an empty vector
(indicating there are no chunks to reverse). This paths in such a way that it
it malloc's a 0-byte buffer (which does allocate bytes on the heap). This buffer
is assigned to a "ChunkedBuffer" instance with the "ChunkedBuffer::set_contiguous"
API. This call fails with a non-ok status and the buffer is leaked.

---

TYPE: BUG
DESC: Fixes a potential memory leak in the filter pipeline