Backward compatibility writes
When a client with a higher version tries to write to an older array
with version greater or equal to 7, we now write in that version. If
the array has a version lesser than 7, we update the array schema
metadata to the current version before the write.

Also added an api to allow the user to migrate an array to the current
version and get the current version from the array schema metadata.

---
TYPE: IMPROVEMENT
DESC: Support backward compat writes