Backward compatibility writes
When a client with a higher version tries to write to an older array
with version greater or equal to 7, we now write in that version. If
the array has a version lesser than 7, we write the fragment in the
current format version.

---
TYPE: IMPROVEMENT
DESC: Support backward compat writes