Add Ubuntu+ASAN to GHA job
This introduces a new workflow for testing with ASAN address sanitization on
Ubuntu. This test knowingly fails. This is OK because its success is not tied
to our precheckin checks.

This mimics a similar job on the Azure Pipelines CI. That job has a false pass.
The job succeeds but ASAN issues are found. Let's add this to GHA so we're
aware of the problems. I will create a follow-up ticket for resolving the
address sanitization issues.