Consolidation: do not re-open array for each fragment
This change removes the need to open an array for read multiple times
during consolidation. It does so first by adding an api to open the
array without fragments and then retreive the list of fragment info
from the storage manager. Then when a fragment is to be consolidated,
we can use the already opened array and set which fragments should be
reloaded, which will only end up hitting the cached fragments.

---
TYPE: IMPROVEMENT
DESC: Consolidation: do not re-open array for each fragment