Adjust Cap'n Proto cmake to better support system installs
Cap'n Proto cmake was adjusted in
[#2100](https://github.com/TileDB-Inc/TileDB/pull/2100) and
[#2210](https://github.com/TileDB-Inc/TileDB/pull/2210) to support
windows and update to Cap'n Proto 0.8.0. This PR builds on those by
making some minor adjustments to better support system installations of
Cap'n Proto and to improve general compatibility. Changes include using
CMAKE_COMMAND variable instead of hardcoding `cmake` as the command. We
now conditionally change the Cap'n Proto compilation command based on if
Cap'n Proto was built as part of the super build or not. Lastly we
remove double importing Cap'n Proto.