Check for minimum validity buffer size (#2261)
Currently, the read path assumes that the validity buffer size is large enough
to contain a validity value for each cell. When this is not true, the core will
overflow the user-set validity buffer.

This adds a runtime check that is invoked when a query is submitted. This ensures
that the validity buffer is large enough.

---

TYPE: IMPROVEMENT
DESC: Runtime check for minimum validity buffer size

Co-authored-by: Joe Maley <joe@tiledb.com>