Subarray partitioner: fixes to hilbert partitioning
This fixes a few issues in the Hilbert partitioning. First, when a range
includes only two Hilbert values, it splits on the left one. Second, it
fixes an int underflow when normalizing and the left range value
coincides with a power of two. Third, it fixes normalization to use the
proper splitting offset. Finally, when reverting from a Hilbert
coordinate value, it now returns the correct maximum possible value to
still fall into that bucket.

---
TYPE: IMPROVEMENT
DESC: Hilbert partitioning fixes