Deprecate tiledb_array_get_timestamp + fix
The end timestamp for an array open is now interpretted as the current time.
I've deprecated the related `tiledb_array_get_timestamp` API because it is no
longer applicable to the start/end timestamp.

I also reverted the C++ API Array::timestamp_end to Array::timestamp. We haven't
shipped this change so it doesn't break anything.

---

TYPE: NO_HISTORY
DESC: Using `NO_HISTORY` since this is a bug in an unreleased feature