Fix Dimension::set_tile_extent for string+hilbert dimension (#2288)
The Dimension::set_tile_extent() implementation errors out if the domain is
unset. We expect the domain to be unset for string dimensions. This patch
allows an empty tile extent to be set on string dimensions. This can be pathed
to by the following path with a hilbert cell-ordered array.
```
Dimension::set_tile_extent
Domain::init
ArraySchema::init
ArraySchema::deserialize
StorageManager::load_array_schema
```

---

TYPE: BUG
DESC: Fix false error message when opening an array with a string dimension and hilbert cell order

Co-authored-by: Joe Maley <joe@tiledb.com>