Fixing tile extent calculations for signed integer domains
In python, users can create a domain for time dimensions like this:
Dim(name='a', domain=(None, None), dtype="datetime64[ms]")

This will create a signed int64 dimension with a domain of min_int and
max_int, with a tile extend covering the whole domain. We used to store
that information and treat it as a signed int64, which would result in
-1. Since lofs of customers might have arrays created like this already,
this fix is treating the tile extent as an unsigned value.

---
TYPE: IMPROVEMENT
DESC: Fixing tile extent calculations for signed integer domains