Handle extra offset w/serialized incomplete query
This adds handling for the extra offset that a user might request (i.e.
for arrow) in serialized queries when there is incomplete queries.

The bug this fixes is that for every incomplete query we were adding n+2
extra offset because the first element of any query results has a 0
offset. This 0th offset is the same value as the "extra" offset included
in the previous query.

The intent of the patch is to handle the fact that only the last part of
the incomplete serialized query needs to include the extra offset. For any
individual query we don't know if it is the last one or not. The last
query in this case doesn't necessarily mean query complete, it could be
the "last" in the sense the user buffers are full and we must let the
user process the partial requests. To work around this issue of not
knowing the last query, we instead remove the first offset of the
current query if the query not the first one run. This works since, as
mentioned, the extra offset is the same as the 0th offset in the n+1
query.