Add serialization support for stats
The stats objects will now we serialized and included their respective
objects. This allows stats to be shipped with the serialized queries for
reads and writes. Stats are not included in the serialization format for
things such as non empty domain, estimated result size or array schema
itself. Stats will only be included for read or write queries that are
serialized.