Initial read refactor
This initial refactor for the readers prepares for the creation of all
new reader classes. To minimize duplicate code, the code for settings
and getting buffers, layout, subarray, config and query condition is
moved to the query. It also creates a new interface, IQueryStrategy,
which all readers and writers will implement and creates a basic class
hierarchy for the strategies. Finally, it creates the reader or writer
dynamically when the query is submitted as a IQueryStrategy.

---
TYPE: IMPROVEMENT
DESC: Initial read refactor