Update spdlog to 1.9.0 (#1973)
* Update spdlog to 1.9.0

Previously we were using spdlog 0.16.3. The older version was not
causing problems with TileDB itself but we no longer played nicely with
system installations of spdlog. Specifically users on
fedora/alpine/arch/gento and osx might have had newer version of spdlog.
There was not API breakage but the internal vs external fmt library was
a problem. The new spdlog has a cmake target variable to declare how the
fmt library should be handled and allows us to now set the proper
compile time definitions needed for the header only library. The newer
library also has proper cmake target.

* Handle spdlog stdout_color_sinks for new version

In the new spdlog 1.9.0 stdout_color_sinks is defined in a separate
header to include. This solves a user reported problem we were
previously unable to reproduce on the older super build spdlog.

Co-authored-by: Luc Rancourt <lucrancourt@gmail.com>
Co-authored-by: Isaiah Norton <ihnorton@users.noreply.github.com>