Sparse global order reader: initial memory budget improvements. (#2413)
Add memory management to Fragment Metadata for rtrees and tile offsets
loading. A budget cap is set on the array by the reader, when we reach
it, it fails the operations. Also keep track of how much data is
currently used for future memory budget calculations.

Freeing rtrees and tile overlap after computing relevant tiles if a
subarray is set.

Move away from using a bitset to determine the next tiles to load and
using a list of tile ranges instead. This is better since for most of
our use cases when having a subarray, the number of tiles is low.

---
TYPE: IMPROVEMENT
DESC: Sparse global order reader: initial memory budget improvements.