Increase default REST request retry count
This increases the number of retries for the REST request from 3 to 25.
With the incremental backoff we do this allows requests to be retried
up to 8 minutes and 47 seconds before erroring. GCS uses a 15 minute
retry by default, AWS also offers a predefined multi-minute retry. The
result will be increased user experience by retrying more and erroring
less.