Remove overly permissive heap API check exception; exclude line from clang-format
This heap check exception in azure.cc was far too permissive. Instead,
exclude the violating line from clang-format to avoid reformatting into
multiple lines (which cannot be matched in find_heap_api_violations
because it works on single lines).

Correct the exclusion to be specific to only the violating signature.