All c_api error paths should reset pointer to null
If we allocate any pointers, we not only need to delete them but we
should reset them to nullptr.

This was highlighted by TileDB-Go where on error we did not see a
nullptr of the original object and our garbage collection attempted to
call the free() function when the structure had already been deleted.
x-ref: https://github.com/TileDB-Inc/TileDB-Go/pull/181