Initial sparse unordered with duplicates reader.
This adds a reader for the sparse unordered with duplicated case. The
reader simply returns all data within the subarray (if set) by
processing all fragments in order. It is hidden behind the same
configuration option as the sparse global order reader but when enabled,
all unit tests are passing.

In this first version, there is simple memory budget management but
serialization is not yet supported.

Also including small fixes to memory management for the sparse global
order reader.

---
TYPE: IMPROVEMENT
DESC: Initial sparse unordered with duplicates reader.